feat: card 를 눌렀을 때 location 이동 완료
카드를 눌러서 모달창이 나타나고 버튼클릭시 이동완료
