+++ trip_location/src/pages/Home.js
@@ -5,86 +5,9 @@ import cityNight from '../images/city_night.jpg';
 import building from '../images/building.jpg';
 import SearchIcon from '@mui/icons-material/Search';
 import PopularDestinations from '../components/contents/main/PopularDestinations';
-import Nav from "../components/Nav/Nav";
 import styled from "@emotion/styled";
 import LocationCard from "../components/contents/main/LocationCard";
-import busan from '../images/haeundae.jpg';
-import jeju from '../images/forest-jeju.jpg';
-import yeosu from '../images/yeosu.jpg';
-import gyeongju from '../images/gyeongju.jpg';
-import gangneung from '../images/gangneung.jpg';
-import seoul from '../images/seoul.jpg';
-import yongin from '../images/yongin.jpg';
-import seosan from '../images/seosan.jpg';
-import jeonju from '../images/jeonju.jpg';
-import ulsan from '../images/ulsan.jpg';
-import mokpo from '../images/mokpo.jpg';
-import suncheon from '../images/suncheon.jpg';
-import Modal from "../components/contents/Modal/Modal";
 
-const cardData = [
-    {
-        image: jeju,
-        title: 'Jeju',
-        description: '제주도',
-    },
-    {
-        image: busan,
-        title: 'Busan',
-        description: '부산',
-    },
-    {
-        image: yeosu,
-        title: 'Yeosu',
-        description: '여수',
-    },
-    {
-        image: gyeongju,
-        title: 'Gyeongju',
-        description: '경주',
-    },
-    {
-        image: gangneung,
-        title: 'Gangneung',
-        description: '강릉',
-    },
-    {
-        image: seoul,
-        title: 'Seoul',
-        description: '서울',
-    },
-    {
-        image: yongin,
-        title: 'Yongin',
-        description: '용인',
-    },
-    {
-        image: seosan,
-        title: 'Seosan',
-        description: '서산',
-    },
-    {
-        image: jeonju,
-        title: 'Jeonju',
-        description: '전주',
-    },
-    {
-        image: ulsan,
-        title: 'Ulsan',
-        description: '울산',
-    },
-    {
-        image: mokpo,
-        title: 'Mokpo',
-        description: '목포',
-    },
-    {
-        image: suncheon,
-        title: 'Suncheon',
-        description: '순천',
-    },
-    // ... 추가 카드 데이터
-];
 
 const MainWrapper = styled.main`
   scroll-snap-type: y mandatory;
@@ -120,29 +43,6 @@ const StyledPaper = styled(Paper)`
   object-fit: cover;
 `;
 
-const contents = css`
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: center;
-  margin: 20px auto;
-  min-width: 80%;
-`;
-
-const searchContainer = css`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  margin-top: 20px;
-  margin-bottom: 20px;
-`;
-
-const searchField = css`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  padding: 5px;
-`;
-
 const StyleTitleTypography = styled(Typography)`
   margin-top: 40px;
   margin-bottom: 10px;
@@ -155,25 +55,9 @@ const StyleSubTitleTypography = styled(Typography)`
   text-align: center;
 `;
 
-const StyleInput = styled(TextField)`
-  margin-bottom: 50px;
-  width: 100%;
-`;
-
 const Home = () => {
     const images = useMemo(() => [cityNight, building], []);
     const [currentImage, setCurrentImage] = useState(images[0]);
-    const [modalOpen, setModalOpen] = useState(false);
-    const [selectedLocation, setSelectedLocation] = useState(null);
-
-    const handleCardClick = (location) => {
-        setSelectedLocation(location);
-        setModalOpen(true);
-    };
-
-    const handleCloseModal = () => {
-        setModalOpen(false);
-    };
 
     useEffect(() => {
         const interval = setInterval(() => {
@@ -204,45 +88,20 @@ const Home = () => {
                 </Container>
             </SectionWrapper>
             <SectionWrapper>
-                <Container>
-                    <div css={searchContainer}>
-                        <StyleTitleTypography  variant="h4" component="h2">
-                            어디로 여행을 떠나시나요?
-                        </StyleTitleTypography>
-                        <StyleSubTitleTypography variant="subtitle1" component="p">
-                            여행지를 검색하거나 목록에서 직접 선택해주세요.
-                        </StyleSubTitleTypography>
-                        <Grid container justifyContent="center" alignItems="center" css={searchField}>
-                            <Grid item xs={12} sm={8} md={6} lg={4}>
-                                <StyleInput
-                                    placeholder="여행지 검색"
-                                    InputProps={{
-                                        startAdornment: <SearchIcon />,
-                                    }}
-                                />
-                            </Grid>
-                        </Grid>
-                    </div>
-                    <Grid container spacing={4} css={contents}>
-                        {cardData.map((data, index) => (
-                            <Grid key={index} item xs={12} sm={6} md={3}>
-                                <LocationCard
-                                    image={data.image}
-                                    title={data.title}
-                                    description={data.description}
-                                    onClick={() => handleCardClick(data)}
-                                />
-                            </Grid>
-                        ))}
-                    </Grid>
-                </Container>
+                <StyleTitleTypography  variant="h4" component="h2">
+                    어디로 여행을 떠나시나요?
+                </StyleTitleTypography>
+                <StyleSubTitleTypography variant="subtitle1" component="p">
+                    여행지를 검색하거나 목록에서 직접 선택해주세요.
+                </StyleSubTitleTypography>
+                <LocationCard/>
             </SectionWrapper>
             <SectionWrapper>
                 <footer>
 
                 </footer>
             </SectionWrapper>
-            <Modal isOpen={modalOpen} onClose={handleCloseModal} destination={selectedLocation} />
+
         </MainWrapper>
     );
 };
